Fixed issue 7, restructured filenames
