+++ src/disk.h
@@ -2,8 +2,6 @@
 #define _NNOFS_DISK_H_
 #include <stdio.h>
 #include <stdlib.h>
-#include "nanofsdisk.h"
-#include "nanofs.h"
 
 typedef FILE* NanoFSDisk;
