+++ src/globalsettings.h
@@ -1,0 +1,12 @@
+#ifndef _NANOFS_GLOBAL_SETTINGS_H_
+#define _NANOFS_GLOBAL_SETTINGS_H_
+
+#define BLOCK_SIZE 4096
+#define MAX_DISK_SIZE (1024*1024*1024)
+#define BLOCKS_COUNT (MAX_DISK_SIZE / BLOCK_SIZE)
+#define BLOCK_FREE_LIST_DIMENSION (BLOCKS_COUNT / 8)
+
+#define MAX_FILENAME 8
+
+
+#endif
