--- src/nanofsglobalsettings.h
@@ -1,11 +1,0 @@
-#ifndef _NANOFS_GLOBAL_SETTINGS_H_
-#define _NANOFS_GLOBAL_SETTINGS_H_
-
-#define BLOCK_SIZE 4096
-#define MAX_DISK_SIZE 1073741824
-#define BLOCK_FREE_LIST_DIMENSTION ((MAX_DISK_SIZE / BLOCK_SIZE) / 256)
-
-#define MAX_FILENAME 8
-
-
-#endif
